+++ src/gwt/src/org/rstudio/studio/client/application/Application.java
@@ -986,7 +986,6 @@ public class Application implements ApplicationEventHandlers
 
       if (!sessionInfo.getAllowShell())
       {
-         commands_.showShellDialog().remove();
          removeTerminalCommands();
       }