+++ src/gwt/src/org/rstudio/studio/client/workbench/commands/Commands.java
@@ -469,7 +469,6 @@ public abstract class
    public abstract AppCommand openProfileInBrowser();
 
    // Tools
-   public abstract AppCommand showShellDialog();
    public abstract AppCommand macPreferences();
    public abstract AppCommand showOptions();
    public abstract AppCommand showCodeOptions();