+++ src/gwt/src/org/rstudio/studio/client/workbench/views/vcs/svn/SVNPane.java
@@ -85,7 +85,7 @@ public class SVNPane extends WorkbenchPane implements Display
       moreMenu.addSeparator();
       moreMenu.addItem(commands_.vcsShowHistory().createMenuItem(false));
       moreMenu.addSeparator();
-      moreMenu.addItem(commands_.showShellDialog().createMenuItem(false));
+      moreMenu.addItem(commands_.newTerminal().createMenuItem(false));
 
       toolbar.addLeftWidget(new ToolbarMenuButton(
           constants_.moreCapitalized(),