+++ NEWS.md
@@ -1,2 +1,6 @@
 
 ## RStudio 2022-10.0 "Elsbeth Geranium" Release Notes
+
+### Deprecated / Removed
+
+- Removed the Tools / Shell command (#11253)